Refactor data structures (#47)
