+++ biogtr/datasets/base_dataset.py
@@ -1,7 +1,7 @@
 """Module containing logic for loading datasets."""
 
 from biogtr.datasets import data_utils
-from biogtr.data_structures import Frame
+from biogtr.io.frame import Frame
 from torch.utils.data import Dataset
 from typing import List, Union
 import numpy as np
