+++ biogtr/datasets/sleap_dataset.py
@@ -7,7 +7,8 @@ import numpy as np
 import sleap_io as sio
 import random
 import warnings
-from biogtr.data_structures import Frame, Instance
+from biogtr.io.frame import Frame
+from biogtr.io.instance import Instance
 from biogtr.datasets import data_utils
 from biogtr.datasets.base_dataset import BaseDataset
 from torchvision.transforms import functional as tvf
