+++ tests/test_training.py
@@ -3,12 +3,13 @@
 import os
 import pytest
 import torch
-from biogtr.data_structures import Frame, Instance
+from biogtr.io.frame import Frame
+from biogtr.io.instance import Instance
 from biogtr.training.losses import AssoLoss
 from biogtr.models.gtr_runner import GTRRunner
 from biogtr.models.global_tracking_transformer import GlobalTrackingTransformer
 from omegaconf import OmegaConf, DictConfig
-from biogtr.config import Config
+from biogtr.io.config import Config
 from biogtr.training.train import main
 
 # TODO: add named tensor tests
